+++ Learning2DEngine/Learning2DEngine/ParticleSimulator/ParticleRenderer.h
@@ -26,13 +26,13 @@ namespace Learning2DEngine
 			GLuint ebo;
 			GLuint vboStatic;
 			GLuint vboDynamic;
-			unsigned int maxObjectSize;
+			size_t maxObjectSize;
 
 			//The first int is the layer
 			//The vector contains the tuples, which renderable at the same time.  
 			//The tuple contains a map, whose keys are the texture ids and values the vector of SpriteRenderData*,
 			//a bool which means, that it uses blend or not, the blend function factor and the max active particle count.
-			std::map<int, std::vector<std::tuple<std::map<GLuint, std::vector<ParticleRenderData*>>, bool, Render::BlendFuncFactor, int>>> particleRenderData;
+			std::map<int, std::vector<std::tuple<std::map<GLuint, std::vector<ParticleRenderData*>>, bool, Render::BlendFuncFactor, size_t>>> particleRenderData;
 			Render::MultiSpriteDynamicData* dynamicData;
 
 			ParticleRenderer();
